refactor: fixup typescript errors
