+++ packages/calcite-components/src/components/dialog/dialog.tsx
@@ -43,7 +43,7 @@ import {
 import { componentOnReady } from "../../utils/component";
 import { SLOTS as PANEL_SLOTS } from "../panel/resources";
 import { HeadingLevel } from "../functional/Heading";
-import { OverlayPositioning } from "../../components";
+import type { OverlayPositioning } from "../../utils/floating-ui";
 import { DialogMessages } from "./assets/dialog/t9n";
 import {
   CSS,
